refactor(ble): split the node-ble scanAndReadRaw into named stages
428 lines down to 212, and scan.ts from 767 to 534. The stages live in
scan-stages.ts; ensureBonded and acquireGattServer stay put, because two test
files import them from scan.js by path.

Seven stages, each keeping the workaround comment that explains the BlueZ bug
it works around:

  acquireBluezAdapter      getAdapter with the stale-connection retry
  waitForTargetDevice      waitDevice under an abort listener that is always
                           removed, or continuous mode leaks one per cycle
  resolvePreConnectAdapter the snapshot and the pre-connect match
  resolveAfterConnect      the characteristic-discovery retry loop
  buildCharMapWithRetry    the second retry loop
  readWithTimeouts         the idle deadline and the absolute session cap
  classifyBleFailure       the #213 watchdog classification
  teardownSession          the whole finally body

Three boundaries are not where they look like they should be, and each is
called out in the code:

  - acquireBluezAdapter deliberately excludes the isPowered check that follows
    it. probeAdapter is assigned between the two, and the #213 classification
    reads it: folding the check in would leave probeAdapter unset when an
    unpowered adapter throws, turning an idle cycle into a wedge-suspect that
    counts toward the watchdog.

  - resolvePreConnectAdapter returns three values, not just the adapter. The
    advertisement snapshot is spread into the POST-connect match as well, and
    dropping it there is exactly the #280 / #318 regression where a dozen
    adapters that key on a company id could never match on Linux.

  - resolveAfterConnect returns the adapter instead of assigning the caller's
    local, and discards the char map it builds. The reading uses a fresh one
    built after the second acquireGattServer, whose adapter may differ.

The order tests from the previous commit are what make this checkable: they
pin the sequence and the two doubled calls, and they were verified to go red on
a swap before any of this moved.

Closes #368
